The progect is done. Will need to doublecheck the requirements and submit.

+++ srcs/requirements/mariadb/tools/config_mariadb.sh
@@ -1,50 +1,45 @@
 #!bin/sh
 echo "..Running 'config_mariadb.sh' script.."
 
-if [ ! -d "/run/mysqld" ]; then
-	mkdir -p /run/mysqld
-	chown -R mysql:mysql /run/mysqld
-fi
-
+# CHeck if mariadb is already installed
 if [ ! -d "/var/lib/mysql/mysql" ]; then
-
+	# creating necessary directories to install and run mariadb
+	mkdir -p /var/lib/mysql /run/mysqld
 	chown -R mysql:mysql /var/lib/mysql /run/mysqld
 
-	mysql_install_db --basedir=/usr --user=mysql --datadir=/var/lib/mysql --rpm > /dev/null
+	# Setting up mariadb configuration for network access
+	sed -i "s|skip-networking|# skip-networking|g" /etc/my.cnf.d/mariadb-server.cnf
+	sed -i "s|.*bind-address\s*=.*|bind-address=0.0.0.0|g" /etc/my.cnf.d/mariadb-server.cnf
 
-	tfile=`mktemp`
-	if [ ! -f "$tfile" ]; then
-		return 1
-	fi
-
-	cat << EOF > $tfile
-USE mysql;
-FLUSH PRIVILEGES;
-
-DELETE FROM mysql.user WHERE User='';
-DROP DATABASE IF EXISTS test;
-DELETE FROM mysql.db WHERE Db='test' OR Db='test\\_%';
-DELETE FROM mysql.user WHERE User='root' AND Host NOT IN ('localhost', '127.0.0.1', '::1');
-
-ALTER USER 'root'@'localhost' IDENTIFIED BY '${MARIADB_ROOT_PASSWORD}';
-
-CREATE DATABASE IF NOT EXISTS ${MARIADB_NAME} CHARACTER SET utf8 COLLATE utf8_general_ci;
-CREATE USER IF NOT EXISTS '${MARIADB_USER}'@'%' IDENTIFIED BY '${MARIADB_PASSWORD}';
-GRANT ALL PRIVILEGES ON ${MARIADB_NAME}.* TO '${MARIADB_USER}'@'%';
-
-FLUSH PRIVILEGES;
-EOF
-
-	/usr/bin/mysqld --user=mysql --bootstrap --verbose=0 < $tfile
-	rm -f $tfile
-
+	# echo "..Installing mariadb.."
+	# mariadb-install-db --user=mysql --datadir=/var/lib/mysql --skip-test-db --skip-name-resolve --auth-root-authentication-method=normal
+	mariadb-install-db --user=mysql --datadir=/var/lib/mysql --skip-test-db
+else
+	echo "..mariadb already installed.."
 fi
 
-# Allowing remote connections
-sed -i "s|skip-networking|# skip-networking|g" /etc/my.cnf.d/mariadb-server.cnf
-sed -i "s|.*bind-address\s*=.*|bind-address=0.0.0.0|g" /etc/my.cnf.d/mariadb-server.cnf
 
-echo "..starting mariadb, port:3306."
+# Check if database already exists
+if [ ! -d "/var/lib/mysql/${DB_NAME}" ]; then
 
-# Starting MariaDB
-exec /usr/bin/mysqld --user=mysql --console
+	cat << EOF > /tmp/setup.sql
+		USE mysql;
+		FLUSH PRIVILEGES;
+		DELETE FROM mysql.user WHERE User='';
+		DELETE FROM mysql.user WHERE User='root' AND Host NOT IN ('localhost', '127.0.0.1', '::1');
+		ALTER USER 'root'@'localhost' IDENTIFIED BY '${DB_ROOT_PASSWORD}';
+		CREATE DATABASE IF NOT EXISTS ${DB_NAME} CHARACTER SET utf8 COLLATE utf8_general_ci;
+		CREATE USER IF NOT EXISTS '${DB_USER}'@'%' IDENTIFIED BY '${DB_PASSWORD}';
+		GRANT ALL PRIVILEGES ON ${DB_NAME}.* TO '${DB_USER}'@'%';
+		FLUSH PRIVILEGES;
+EOF
+	# Starting mariadb in a bootstrap mode to create the database and user
+	mariadbd --user=mysql --bootstrap --verbose=0 < /tmp/setup.sql
+	rm -f /tmp/setup.sql
+
+	echo "..database '${DB_NAME}' created.."
+else
+	echo "..database '${DB_NAME}' already exists.."
+fi
+
+exec "$@"
